+++ packages/orq-rc/src/models/errors/index.ts
@@ -8,8 +8,6 @@ export * from "./createeval.js";
 export * from "./deleteagent.js";
 export * from "./deletecontact.js";
 export * from "./deleteeval.js";
-export * from "./duplicateagent.js";
-export * from "./duplicatetool.js";
 export * from "./getagent.js";
 export * from "./getagenttask.js";
 export * from "./getevals.js";