+++ packages/orq-rc/RELEASES.md
@@ -5149,3 +5149,13 @@ Based on:
 - [typescript v4.0.0-rc.49] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.49 - packages/orq-rc
+
+## 2025-11-19 11:31:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.50] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.50 - packages/orq-rc